refactor: drop the raw last_checkpoint_hint getter and group hint tests in a file
The retained hint is now the pub(crate) last_checkpoint_metadata field (incremental reads it directly); checkpoint_hint() -- the identity-gated accessor -- is the only hint method, removing the last_checkpoint_hint/checkpoint_hint name clash. Move the log_segment hint tests into log_segment/tests/checkpoint_hint.rs.
